# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API VisionHub360",
        "description": "API RESTful con JWT y Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/suppliers": {
            "post": {
                "tags": [
                    "Gestión de Proveedores"
                ],
                "summary": "Registrar un nuevo proveedor",
                "operationId": "4a973a8e1c3cae75708d25833f0af09e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Joyas del Sur"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "12345678-9"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "contacto@gafasygafas.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "98787 5656"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Av. Central 456"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Colombia"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Bogotá"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proveedor creado correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/suppliers/{supplier}": {
            "put": {
                "tags": [
                    "Gestión de Proveedores"
                ],
                "summary": "Actualizar los datos de un proveedor existente",
                "operationId": "d09c0fa0576183994a045bd472a86742",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "description": "ID del proveedor a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Joyas del Norte"
                                    },
                                    "document": {
                                        "type": "string",
                                        "example": "98765432-1"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "contacto@joyasnorte.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "320 987 6543"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Cra 45 #22-18"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Medellín"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Colombia"
                                    },
                                    "active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proveedor actualizado correctamente"
                    },
                    "404": {
                        "description": "Proveedor no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Gestión de Proveedores"
                ],
                "summary": "Eliminar un proveedor (soft delete)",
                "operationId": "1eba351ef4ff602e5a63c6110e65ba9e",
                "parameters": [
                    {
                        "name": "supplier",
                        "in": "path",
                        "description": "ID del proveedor a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proveedor eliminado correctamente"
                    },
                    "404": {
                        "description": "Proveedor no encontrado"
                    },
                    "403": {
                        "description": "No tiene permisos para eliminar este proveedor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Gestión de Proveedores"
                ],
                "summary": "Lista y busca proveedores con paginación",
                "operationId": "50ff71ddda34eeb1f112adf728bdbf36",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de proveedores por página. Por defecto es 10.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página a recuperar.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda para filtrar por nombre, documento, email o teléfono.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Lista de proveedores con metadatos de paginación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de proveedores obtenida correctamente"
                                        },
                                        "count": {
                                            "description": "Conteo total de registros en todas las páginas.",
                                            "type": "integer",
                                            "example": 30
                                        },
                                        "data": {
                                            "properties": {
                                                "suppliers": {
                                                    "description": "Objeto de paginación de Laravel",
                                                    "properties": {
                                                        "current_page": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "description": "Detalle del objeto Proveedor (ejemplo de campos)",
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 1
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Proveedor Mayorista S.A."
                                                                    },
                                                                    "document_number": {
                                                                        "type": "string",
                                                                        "example": "900123456-7"
                                                                    },
                                                                    "email": {
                                                                        "type": "string",
                                                                        "example": "contacto@proveedor.com"
                                                                    },
                                                                    "phone": {
                                                                        "type": "string",
                                                                        "example": "6015551234"
                                                                    },
                                                                    "created_at": {
                                                                        "type": "string",
                                                                        "format": "date-time"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "last_page": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "per_page": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "total": {
                                                            "type": "integer",
                                                            "example": 30
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}